Update to new release version of Discord4J
This update is minor but does future proof with easier version updating for Discord4J. Also works better with the MavenCentral/JCenter releases. (Note: I will be updating this to the 3.x version of Discord4J upon release. v3 is currently in Beta on the Discord4J side and all features are not in it yet so will not be working on it.)